backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
firebase-admin>=6.2.0
python-multipart>=0.0.6
shortuuid>=1.0.11
pydantic>=2.4.2
email-validator>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
reportlab>=3.6.0
python-docx>=0.8.11
chardet>=4.0.0
google-cloud-vision>=3.4.0
google-cloud-speech>=2.20.0
pillow>=10.0.0
numpy>=1.23.0
google-generativeai>=0.8.4
SpeechRecognition>=3.10.0
moviepy>=1.0.3
nltk>=3.8.1
document
google-cloud-documentai>=2.20.1
google-cloud-speech
ffmpeg-python>=0.2.0
PyPDF2>=3.0.0
langchain>=0.0.267
pandas>=2.1.0
scikit-learn>=1.3.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.1
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.1
pre-commit>=3.4.0
aiofiles>=23.2.1
torch
transformers>=4.30.0
google-cloud-language>=2.9.1
librosa
transformers
pytz
python-Levenshtein>=0.12.0
mediapipe
opencv-python
requests
google-cloud-storage
protobuf==4.25.7
sentencepiece
pydub
huggingface_hub
transformers
accelerate
bitsandbytes
sentencepiece
google-cloud-videointelligence==2.16.1